Adding rebuild to scripts

+++ scripts/build-macosx.sh
@@ -2,8 +2,25 @@
 DIR="$(cd -P "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 cd "$DIR/.." || exit 1
 
-rm -rf build
-bash scripts/build.sh
+rebuild=true
+
+while [ $# -gt 0 ]; do
+    case "$1" in
+        --no-rebuild)
+            rebuild=false
+            ;;
+        *)
+            box "Error: Invalid argument."
+            exit 1
+    esac
+    shift
+done
+
+if [ "$rebuild" == "true" ]; then
+    rm -rf build
+    bash scripts/build.sh
+fi
+
 cd build/runtime-pyside6 || exit 1
 
 rm $(find . -name "*.pyc")
@@ -30,7 +47,7 @@ cp ../../assets/native/mac/Info.plist dist/LogarithmPlotter.app/Contents/Info.pl
 
 # Remove QtWebEngine, 3D and all other unused libs libs
 rm -rf dist/LogarithmPlotter.app/Contents/MacOS/{QtWeb*,*3D*,QtRemote*,QtPdf,QtCharts,QtLocation,QtTest,QtMultimedia,QtSpatialAudio,QtDataVisualization,QtQuickParticles,QtChartsQml,QtScxml,QtDataVisualizationQml,QtTest,QtPositioningQuick,QtQuickTest,QtSql,QtSensorsQuick}
-rm -rf dist/LogarithmPlotter.app/Contents/MacOS/PySide6/{QtNetwork.abi3.so}
+rm -rf dist/LogarithmPlotter.app/Contents/MacOS/PySide6/QtNetwork.abi3.so
 
 # Removing QtQuick3D
 rm -rf dist/LogarithmPlotter.app/Contents/MacOS/PySide6/Qt/qml/{QtQuick3D,Qt3D,QtWebEngine}
