+++ scripts/run-tests.sh
@@ -1,8 +1,26 @@
 #!/bin/bash
-cd "$(dirname "$(readlink -f "$0" || realpath "$0")")/.."
+cd "$(dirname "$(readlink -f "$0" || realpath "$0")")/.." || exit 1
 
-rm -rf build
-bash scripts/build.sh
+rebuild=true
+
+while [ $# -gt 0 ]; do
+    case "$1" in
+        --no-rebuild)
+            rebuild=false
+            ;;
+        *)
+            box "Error: Invalid argument."
+            exit 1
+    esac
+    shift
+done
+
+if [ "$rebuild" == "true" ]; then
+    rm -rf build
+    bash scripts/build.sh
+fi
+
+
 
 # Run python tests
 cp -r runtime-pyside6/tests build/runtime-pyside6
@@ -12,6 +30,6 @@ PYTHONPATH="$PYTHONPATH:." pytest --cov=LogarithmPlotter --cov-report term-missi
 cd ../../
 
 # Run js tests
-cd common
+cd common || exit 1
 npm test
